libcamera: pipeline: uvcvideo: Add V4L2 device object representing uvc metadata video node
Identify and open the UVC header's video node. This will ultimately gives us access to metadata associated with the video stream of the uvc camera.  The user will not have access to this video node and will not need to manage its data in any way.

The node is identified as the sole remaining video node entity associated with a provided media device, after the video node for the sensor has already been found.  If this node does not have metadata capabilites, initialization fails and the node is not opened.

For now, a found and successfully opened metadata video node results in an INFO log reporting the success.

Signed-off-by: Gabby George <gabbymg94@gmail.com>