libcamera: Make file descriptor getter for a v4l2 device public.
Currently, the file descriptor getter associated with a V4L2 device is
a private function.  Objects that use V4L2 devices may want to get access
to the file descriptor.  The file descriptor may be needed to perform mmap
calls on buffers which are not exported via dmabuf handles.

Make this getter public.

Signed-off-by: Gabby George <gabbymg94@gmail.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>