libcamera: pipeline: uvcvideo: Allocate and queue UVC metadata stream buffers to be filled by device. Show it works with print statements.
To get more accurate UVC timestamps, libcamera needs access to the UVC metadata stream,
if available.  Within the UVC pipeline, allocate buffers for UVC metadata and map them
into libcamera's virtual memory space.  Start the metadata buffer stream, and register
a bufferReady callback so the uvc pipeline can process the information contained in
the metadata buffers.

Queue and dequeue these buffers, and then use the frame buffer's
sequence field to match the metadata buffers with a video buffer
with the same sequence. When a match is found, print a debug
statement showing so.

Signed-off-by: Gabby George <gabbymg94@gmail.com>