libcamera: MappedFrameBuffer: Use stored plane offset
The MappedFrameBuffer causes a failure in mmap when
attempting to map UVC metadata planes.  UVC metadata (four character
code UVCH) does not support exporting buffers as file descriptors, so
mmap can be used to give libcamera access to the metadata buffer
planes. It is convenient to use the already-existing
MappedFrameBuffers class, but the class must be modified to support
mapping using file descriptors of the video node itself.

To do this, mmap needs information obtained through a call to
QUERYBUF, namely, the plane offset for buffer planes. Modify the
constructor of a MappedFrameBuffer to use the plane offset directly in
the call to mmap, rather than the hard-coded 0 value.

The current version does not work when a buffer cannot be exported as
a dma buf fd. The fd argument to mmap must be the one obtained on a
call to open() for the video node (ie, /dev/videoX). This method of
mapping buffer planes requires the arguments to mmap to be exactly the
length and offset QUERYBUF provides.  mmap will return a -EINVAL if
this is not the case.

Signed-off-by: Gabby George <gabbymg94@gmail.com>