libcamera: v4l2 device: Store buffer info in planes
To perform a memory mapping using mmap, the MappedFrameBuffer class
needs the plane offset and file descriptor information of the frame
buffer's plane(s). This information is provided in the response to
REQBUF, which happens during buffer allocation. Store the plane offset
and file descriptor information in the buffer's plane at the time of
allocation.

Currently, there is a metadata buffer type (metadata format UVCH) that
does not support exporting buffers using EXPBUF, so this should only
be done if the buffer type is metadata capture.

Signed-off-by: Gabby George <gabbymg94@gmail.com>