libcamera: pipeline: uvcvideo: Add circular buffer for timestamps
To calculate UVC timestamps, metadata from some number of frames (32,
here) back is needed.  To store this information, use a circular
buffer.  This makes it quick and readable to remove the oldest item
and add a new one. It also gives fast access to the front and back of
the queue, the data of which are both needed at once.

Signed-off-by: Gabby George <gabbymg94@gmail.com>