libcamera: pipeline: uvcvideo: Store time metadata
Parse the raw timing data out of the metadata packets and store each
sample in a circular buffer of size 32. To convert the latest frame's
metadata into a system clock timestamp, use the metadata time
information from 32 frames back stored in the back of the circular
buffer.

For every metadata buffer that comes in, store the timing information
in the circular queue. The information at the head of the queue may be
used instantly in the case where a corresponding video buffer has come
in, or it may be used when the video buffer eventually does arrive. In
any case, the information at the head of the queue will be used when
it comes time for the timestamp to be set in the request metadata for
the video buffer.

If there has only been one frame of metadata, use the default (video
buffer-provided) timestamp.  If the circular buffer has not yet been
filled with 32 frames, use the oldest frame. When the circular buffer
reaches 32 frames, remove the oldest item.

This approach creates synchronization between the metadata buffer that
will be used to update the video buffer's requests' timestamp and the
timing data itself, without having to perform checks ensuring
synchronization. It also creates easy and efficient access to the
front and back of the circular buffer.

Signed-off-by: Gabby George <gabbymg94@gmail.com>